Test the entity subscriptions api
